pyproject.toml
[tool.poetry]
name = "my-hackathon-project"
version = "0.1.0"
description = "Project with specified dependencies"
authors = ["Taowei Ji"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
tenacity = "9.0.0"
langchain = "0.3.12"
langchain-openai = "0.2.12"
langchain_community = "0.3.12"
langgraph = "0.2.59"
langchain_chroma = "0.1.4"
pandas = "2.2.3"
pypdf = "5.1.0"
nbformat = "5.10.4"
azure-core = "^1.32.0"
faiss-cpu = "^1.10.0"
scikit-learn = "^1.6.1"
sentence-transformers = "3.4.1"

[tool.poetry.dev-dependencies]
# Add any development/testing libraries here

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"